+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this package will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.0.1] - 2021-08-26
+
+### Changed
+  - Minor update to the documentation 
+  - Unity minimum version is now 2020.3.16f1
+  
+### Fixed
+  - Shader related errors on Unity 2021.2
+
 ## [1.0.1-pre.585] - 2021-07-27
 
 ### Changed